+++ src/scenes/menu-scene.js
@@ -2,6 +2,7 @@ import { Container } from '../engine.js';
 import { STATES } from '../state-machine.js';
 import { ASSETS, CreateSprite } from '../sprites.js';
 import { showNotification } from '../overlay.js'
+import { toggleFullScreenButton } from './common.js';
 
 export function createMenuScene({ stateMachine }) {
     const container = new Container();
@@ -44,6 +45,8 @@ export function createMenuScene({ stateMachine }) {
             action: () => showNotification({ title: 'Here will be settings.', parent: container }),
             position: { x: 420, y: 355 },
         });
+
+        container.addChild(toggleFullScreenButton);
     }
 
     return {